Streamline rebuild scripts
